开始引入JIT，目前使用LLVM官方的JIT模板

+++ ast/const/int32.cpp
@@ -4,7 +4,7 @@
 namespace AST {
     namespace Const {
         llvm::Value *TInt32::genCode() {
-            return llvm::ConstantInt::get(module->context, llvm::APInt(32, uint64_t(data), true));
+            return llvm::ConstantInt::get(module->context, llvm::APInt(32, data, true));
         }
     }
 }
